Update .gitignore, CHANGELOG, and project metadata for v1.0 release
- Enhanced .gitignore to include specific environment directories and versioned golden baseline JSON files.
- Updated CHANGELOG with detailed entries for new features, breaking changes, and improvements in the v1.0 release, including the introduction of a layered public API and various domain-specific enhancements.
- Modified MANIFEST.in and setup.py to reflect changes in file structure and dependencies, ensuring compatibility with the new project layout.
- Adjusted pyproject.toml to update versioning and dependency specifications for improved package management.

This commit marks the transition to version 1.0, consolidating significant architectural changes and feature additions.

+++ tests/domain/test_supervoxel.py
@@ -58,7 +58,7 @@ def test_slic_features_are_per_supervoxel_means() -> None:
     voxel_labels = np.asarray(unit.label_array)[tuple(field.voxel_index.T)]
     expected = (
         pd.DataFrame(field.values, columns=list(field.feature_names))
-        .groupby(voxel_labels)
+        .groupby(voxel_labels.astype(np.int64))
         .mean()
     )
     pd.testing.assert_frame_equal(
